refactor: isolate Parquet Variant normalization
